requirements.txt
streamlit
google-generativeai
mysql-connector-python
pandas